Crate sui_storage

Re-exports

pub use indexes::IndexStore;

pub use indexes::IndexStoreTables;

Modules

blob

http_key_value_store

indexes

IndexStore supports creation of various ancillary indexes of state in SuiDataStore. The main user of this data is the explorer.

key_value_store

Immutable key/value store trait for storing/retrieving transactions, effects, and events to/from a scalable.

key_value_store_metrics

mutex_table

object_store

package_object_cache

sharded_lru

write_path_pending_tx_log

WritePathPendingTransactionLog is used in the transaction write path (e.g. in TransactionOrchestrator) for transaction submission processing. It helps to achieve:

Enums

FileCompression

StorageFormat

Constants

SHA3_BYTES

Functions

compress

compute_sha3_checksum

compute_sha3_checksum_for_bytes

compute_sha3_checksum_for_file

make_iterator

read

verify_checkpoint

verify_checkpoint_range

verify_checkpoint_with_committee