Trait sui_storage::event_store::EventStore

pub trait EventStore {
    fn add_events<'life0, 'life1, 'async_trait>(
        &'life0 self,
        events: &'life1 [EventEnvelope]
    ) -> Pin<Box<dyn Future<Output = Result<u64, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn all_events<'life0, 'async_trait>(
        &'life0 self,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn events_by_transaction<'life0, 'async_trait>(
        &'life0 self,
        digest: TransactionDigest,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn events_by_type<'life0, 'async_trait>(
        &'life0 self,
        event_type: EventType,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn events_by_module_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        module: &'life1 ModuleId,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn events_by_move_event_struct_name<'life0, 'life1, 'async_trait>(
        &'life0 self,
        move_event_struct_name: &'life1 str,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn events_by_sender<'life0, 'life1, 'async_trait>(
        &'life0 self,
        sender: &'life1 SuiAddress,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn events_by_recipient<'life0, 'life1, 'async_trait>(
        &'life0 self,
        recipient: &'life1 Owner,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn events_by_object<'life0, 'life1, 'async_trait>(
        &'life0 self,
        object: &'life1 ObjectID,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn event_iterator<'life0, 'async_trait>(
        &'life0 self,
        start_time: u64,
        end_time: u64,
        cursor: EventID,
        limit: usize,
        descending: bool
    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

An EventStore supports event ingestion and flexible event querying One can think of events as logs. They represent a log of what is happening to Sui. Thus, all different kinds of events fit on a timeline, and one should be able to query for different types of events that happen over that timeline.

Required Methods

fn add_events<'life0, 'life1, 'async_trait>(
&'life0 self,
events: &'life1 [EventEnvelope]
) -> Pin<Box<dyn Future<Output = Result<u64, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Adds a batch of transaction-related events to the EventStore. Semantics:

The batch is appended to the store.

The batch may contain events from multiple transactions.

Each event must have a unique (seq_num, event_num) tuple - events that duplicate this key will be ignored.

Returns Ok(rows_affected).

fn all_events<'life0, 'async_trait>(
&'life0 self,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns at most limit events emitted by all transaction, ordered .

fn events_by_transaction<'life0, 'async_trait>(
&'life0 self,
digest: TransactionDigest,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns at most limit events emitted by a given transaction, sorted in time order defined by the descending parameter.

fn events_by_type<'life0, 'async_trait>(
&'life0 self,
event_type: EventType,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns at most limit events of a certain EventType (e.g. TransferObject) within [start_time, end_time), sorted in time order defined by the descending parameter.

fn events_by_module_id<'life0, 'life1, 'async_trait>(
&'life0 self,
module: &'life1 ModuleId,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns at most limit events emitted in a certain Module ID during sorted in time order defined by the descending parameter.

fn events_by_move_event_struct_name<'life0, 'life1, 'async_trait>(
&'life0 self,
move_event_struct_name: &'life1 str,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns at most limit events with the move event struct name (e.g. 0x2::devnet_nft::MintNFTEvent) emitted sorted in time order defined by the descending parameter.

fn events_by_sender<'life0, 'life1, 'async_trait>(
&'life0 self,
sender: &'life1 SuiAddress,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns at most limit events associated with a certain sender sorted in time order defined by the descending parameter.

fn events_by_recipient<'life0, 'life1, 'async_trait>(
&'life0 self,
recipient: &'life1 Owner,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns at most limit events associated with a certain recipient sorted in time order defined by the descending parameter.

fn events_by_object<'life0, 'life1, 'async_trait>(
&'life0 self,
object: &'life1 ObjectID,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns at most limit events associated with a certain object id sorted in time order defined by the descending parameter.

fn event_iterator<'life0, 'async_trait>(
&'life0 self,
start_time: u64,
end_time: u64,
cursor: EventID,
limit: usize,
descending: bool
) -> Pin<Box<dyn Future<Output = Result<Vec<StoredEvent>, SuiError>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Generic event iterator that returns events emitted between [start_time, end_time), sorted in time order defined by the descending parameter.