Trait sui_storage::write_ahead_log::WriteAheadLog

pub trait WriteAheadLog<'a, C> {
    type Guard: TxGuard<'a>;
    type LockGuard;

    fn begin_tx<'b, 'async_trait>(
        &'a self,
        tx: &'b TransactionDigest,
        cert: &'b C
    ) -> Pin<Box<dyn Future<Output = SuiResult<Self::Guard>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'b: 'async_trait,
        'a: 'async_trait;
    fn acquire_lock<'life0, 'async_trait>(
        &'a self,
        tx: &'life0 TransactionDigest
    ) -> Pin<Box<dyn Future<Output = Self::LockGuard> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'a: 'async_trait,
        'life0: 'async_trait;
    fn read_one_recoverable_tx<'async_trait>(
        &'a self
    ) -> Pin<Box<dyn Future<Output = SuiResult<Option<(C, Self::Guard)>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'a: 'async_trait;
}

Required Associated Types

type Guard: TxGuard<'a>

type LockGuard

Required Methods

fn begin_tx<'b, 'async_trait>(
&'a self,
tx: &'b TransactionDigest,
cert: &'b C
) -> Pin<Box<dyn Future<Output = SuiResult<Self::Guard>> + Send + 'async_trait>>where
Self: 'async_trait,
'b: 'async_trait,
'a: 'async_trait,

Begin a confirmation transaction identified by its digest, with the associated cert.

The possible return values mean:

Ok(None) => There was a concurrent instance of the same tx in progress, but it ended without being committed. The caller may not proceed processing that tx. A TxGuard for that tx can be (eventually) obtained by calling read_one_recoverable_tx().

Ok(Some(TxGuard)) => No other concurrent instance of the same tx is in progress, nor can one start while the guard is held. However, a prior instance of the same tx could have just finished, so the caller may want to check if the tx is already sequenced before proceeding.

Err(e) => An error occurred.

fn acquire_lock<'life0, 'async_trait>(
&'a self,
tx: &'life0 TransactionDigest
) -> Pin<Box<dyn Future<Output = Self::LockGuard> + Send + 'async_trait>>where
Self: 'async_trait,
'a: 'async_trait,
'life0: 'async_trait,

fn read_one_recoverable_tx<'async_trait>(
&'a self
) -> Pin<Box<dyn Future<Output = SuiResult<Option<(C, Self::Guard)>>> + Send + 'async_trait>>where
Self: 'async_trait,
'a: 'async_trait,

Recoverable TXes are TXes that we find in the log at start up (which indicates we crashed while processing them) or implicitly dropped TXes (which can happen because we errored out of the write path and implicitly dropped the TxGuard).

This method pops one recoverable tx from that log, acquires the lock for that tx, and returns a TxGuard.

The caller is responsible for running the tx to completion.

Recoverable TXes will remain in the on-disk log until they are explicitly committed.