Module sui_storage::write_path_pending_tx_log

WritePathPendingTransactionLog is used in the transaction write path (e.g. in TransactionOrchestrator) for transaction submission processing. It helps to achieve:

At one time, a transaction is only processed once.

When Fullnode crashes and restarts, the pending transaction will be loaded and retried.

Structs

WritePathPendingTransactionLog

WritePathPendingTransactionTableConfigurator

Create config structs for configuring DBMap tables

WritePathPendingTransactionTableReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

Type Aliases

IsFirstRecord