sui_storage::object_store
Module util

Structs

Manifest

PerEpochManifest

Constants

MANIFEST_FILENAME

Functions

copy_file

copy_files

copy_recursively

delete_files

delete_recursively

exists

find_all_dirs_with_epoch_prefix

This function will find all child directories in the input store which are of the form “epoch_num” and return a map of epoch number to the directory path

find_all_files_with_epoch_prefix

This function will find all child directories in the input store which are of the form “epoch_num” and return a map of epoch number to the directory path

find_missing_epochs_dirs

This function will find missing epoch directories in the input store and return a list of such epoch numbers. If the highest epoch directory in the store is epoch_N then it is expected that the store will have all epoch directories from epoch_0 to epoch_N. Additionally, any epoch directory should have the passed in marker file present or else that epoch number is already considered as missing

get

get_path

list_all_epochs

path_to_filesystem

put

run_manifest_update_loop

write_snapshot_manifest