sui_storage
Module write_path_pending_tx_log

WritePathPendingTransactionLog is used in the transaction write path (e.g. in TransactionOrchestrator) for transaction submission processing. It helps to achieve:

At one time, a transaction is only processed once.

When Fullnode crashes and restarts, the pending transaction will be loaded and retried.

Structs

WritePathPendingTransactionLog

WritePathPendingTransactionTableReadOnly

Type Aliases

IsFirstRecord